requirements.txt
dotenv
pymongo
python-telegram-bot
python-telegram-bot[webhooks]
google-genai